fix: 取引リストを更新すると上のタブまで再描画されるミスを修正
