+++ client/app/lib/Views/home_view.dart
@@ -3,7 +3,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../Usecases/provider.dart';
 import 'search/search_view.dart';
 import 'user/my_page_view.dart';
 import 'receiver/furniture_list_view.dart';
@@ -31,43 +30,10 @@ class HomeView extends HookConsumerWidget {
       return null;
     }, [camera]);
 
-    final tradeState = ref.watch(tradeListProvider);
-
-    Future<void> reloadTradeList() {
-      // ignore: unused_result
-      ref.refresh(tradeListProvider);
-      return ref.read(tradeListProvider.future);
-    }
-
-    // 画面を移動した時に自動で更新
-    useEffect(() {
-      reloadTradeList();
-      return null;
-    }, [selectedView.value]);
-
     final viewWidgets = [
       FurnitureListView(selectedView: selectedView),
       SearchView(cameraController: cameraController),
-      // 下に引っ張った時に更新
-      RefreshIndicator(
-        onRefresh: () => reloadTradeList(),
-        // 取引リストの取得
-        child: tradeState.when(
-          loading: () => const Center(
-            child: CircularProgressIndicator(),
-          ),
-          error: (error, __) => Center(
-            child: Text('$error'),
-          ),
-          skipLoadingOnRefresh: false,
-          data: (data) {
-            return RegisterProductView(
-              tradeList: data,
-              cameraController: cameraController,
-            );
-          },
-        ),
-      ),
+      RegisterProductView(cameraController: cameraController),
       MyPageView(camera: camera),
     ];
